fix: move useNonce to shared\src\hooks
