+++ shared/src/hooks/chains/ethereum/index.ts
@@ -1,4 +1,4 @@
-export { useNonce } from './useNonce';
+export { useNonce } from '../../useNonce';
 export { useVerifySignature, setTokenSuccessCallback } from './useVerifySignature';
 export { useUserProfile } from './useUserProfile';
 export { usePetsContract } from './usePetsContract';
